fbcli/requirements.txt
ask==0.0.8
prompt-toolkit==2.0.7
Pygments==2.3.0
PyYAML==3.13
logbook==1.5.2
rainbow-logging-handler==2.2.2
paramiko==2.5.0
redis-py-cluster==1.3.6
terminaltables==3.1.0
hiredis==0.2.0
retrying==1.3.3
Werkzeug==0.14.1
Click==7.0
PyHive==0.6.1
thrift==0.11.0
thrift-sasl==0.3.0
fire==0.1.3
